% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-grouped_axis.R
\name{guide_grouped_axis}
\alias{guide_grouped_axis}
\title{Guide that draws the group-wise labels for the x-axis}
\usage{
guide_grouped_axis(title = waiver(), ...)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2:waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[ggplot2:labs]{labs()}} is used for the title.}

\item{...}{passed on to `ggplot2::guide_axis`}
}
\description{
Guide that draws the group-wise labels for the x-axis
}
