% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grouped_axis.R
\name{theme_grouped_axis}
\alias{theme_grouped_axis}
\title{Theme elements specifically for the grouped axis}
\usage{
theme_grouped_axis(
  axis.grouping.line = NULL,
  axis.grouping.line_padding = unit(20, "points"),
  axis.grouping.label = NULL,
  axis.grouping.line_height = unit(20, "points"),
  ...
)
}
\arguments{
\item{axis.grouping.line}{a line element}

\item{axis.grouping.line_padding}{by how much should each line be extended in each direction}

\item{axis.grouping.label}{a text element}

\item{axis.grouping.line_height}{how much space should there be around the line}

\item{...}{additional parameters that are passed on to `theme`}
}
\description{
Theme elements specifically for the grouped axis
}
